Properly cast floats for sqrt.

+++ src/vector.c
@@ -45,6 +45,5 @@ float Vector_dot(Vector a, Vector b) {
 
 
 float Vector_mag(Vector self) {
-    return sqrt(Vector_dot(self, self));
+    return (float) sqrt((double) Vector_dot(self, self));
 }
-
